# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octroi API
  description: API gateway for AI agent tool access.
  version: "0.1.0"
servers:
  - url: /
    description: Default server
...
security: []
paths:
...
  /api/v1/tools/search:
    get:
      operationId: searchTools
      tags: [public]
      summary: Search tools
      parameters:
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Paginated list of matching tools.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Public: List tools ----------
...
  /api/v1/admin/agents:
    post:
      operationId: createAgent
      tags: [admin-agents]
      summary: Create an agent
      description: Creates an agent and generates an API key. The plaintext key is returned only in this response.
      security:
        - AdminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created. Contains the plaintext API key (shown only once).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAgents
      tags: [admin-agents]
      summary: List agents
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryCursor"
      responses:
        "200":
          description: Paginated list of agents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/agents/{id}:
    put:
      operationId: updateAgent
      tags: [admin-agents]
      summary: Update an agent
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathAgentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentInput"
      responses:
        "200":
          description: Agent updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteAgent
      tags: [admin-agents]
      summary: Delete an agent
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathAgentID"
      responses:
        "204":
          description: Agent deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Admin: Budget management ----------
  /api/v1/admin/agents/{agentID}/budgets/{toolID}:
    put:
      operationId: setBudget
      tags: [admin-budgets]
      summary: Set a budget for an agent on a tool
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetAgentID"
        - $ref: "#/components/parameters/PathBudgetToolID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBudgetInput"
      responses:
        "200":
          description: Budget set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: getBudget
      tags: [admin-budgets]
      summary: Get budget for an agent on a tool
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetAgentID"
        - $ref: "#/components/parameters/PathBudgetToolID"
      responses:
        "200":
          description: The budget.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin/agents/{agentID}/budgets:
    get:
      operationId: listBudgets
      tags: [admin-budgets]
      summary: List all budgets for an agent
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/PathBudgetAgentID"
      responses:
        "200":
          description: List of budgets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Admin: Usage queries ----------
...
  /api/v1/agents/me:
    get:
      operationId: getSelfAgent
      tags: [agent]
      summary: Get the authenticated agent
      security:
        - AgentBearer: []
      responses:
        "200":
          description: The authenticated agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Agent-authed: Usage ----------
...
components:
...
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      description: Admin API key passed as a Bearer token.
    AgentBearer:
      type: http
      scheme: bearer
      description: Agent API key passed as a Bearer token.

  # ---------- Reusable parameters ----------
  parameters:
...
    PathAgentID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Agent ID.
    PathBudgetAgentID:
      name: agentID
      in: path
      required: true
      schema:
        type: string
      description: Agent ID.
    PathBudgetToolID:
      name: toolID
      in: path
      required: true
      schema:
        type: string
      description: Tool ID.
...
    QueryQ:
      name: q
      in: query
      schema:
        type: string
      description: Free-text search query.
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
      description: Maximum number of items to return.
    QueryCursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for fetching the next page.
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: tool not found

    # --- Manifest ---
...
    ToolPublic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        auth_type:
          type: string
        tags:
          type: array
          items:
            type: string
        pricing_model:
          type: string
        pricing_amount:
          type: number
          format: double
        pricing_currency:
          type: string
        rate_limit:
          type: integer
        budget_limit:
          type: number
          format: double
        budget_window:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Tool (admin view, includes endpoint and auth_config) ---
...
    ToolListResponse:
      type: object
      required: [tools]
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolPublic"
        next_cursor:
          type: string
          description: Opaque cursor for the next page. Absent when there are no more results.

    # --- Agent ---
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        api_key_prefix:
          type: string
        team:
          type: string
        rate_limit:
          type: integer
        created_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        team:
          type: string
        rate_limit:
          type: integer
    CreateAgentResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        api_key_prefix:
          type: string
        api_key:
          type: string
          description: Plaintext API key. Only returned once at creation time.
        team:
          type: string
        rate_limit:
          type: integer
        created_at:
          type: string
          format: date-time
    UpdateAgentInput:
      type: object
      description: All fields are optional. Only provided fields are applied.
      properties:
        name:
          type: string
        team:
          type: string
        rate_limit:
          type: integer
    AgentListResponse:
      type: object
      required: [agents]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        next_cursor:
          type: string
          description: Opaque cursor for the next page. Absent when there are no more results.

    # --- Budget ---
    Budget:
      type: object
      properties:
        id:
          type: string
        agent_id:
          type: string
        tool_id:
          type: string
        daily_limit:
          type: number
          format: double
        monthly_limit:
          type: number
          format: double
    SetBudgetInput:
      type: object
      properties:
        daily_limit:
          type: number
          format: double
        monthly_limit:
          type: number
          format: double
    BudgetListResponse:
      type: object
      required: [budgets]
      properties:
        budgets:
          type: array
          items:
            $ref: "#/components/schemas/Budget"

    # --- Usage / Metering ---
...